feat: Removed course header stuff
