+++ src/course-tabs/CourseTabsNavigation.jsx
