+++ src/course-tabs/CourseTabsNavigation.test.jsx
